Basic working Handoff engine.
Calls out to the REST service with a different set of headers to
before. Returns useful information when authenticated, returns sensible errors if
not.

Some Doxygen added now, more required.

I am concerned about the blocking behaviour of the authenticator client,
that needs tested - it may be that this is the wrong type of HTTP
client.