Check for an x- header instead. POST.
Instead of using purely query parameters, look for the x-rgw-storequery header
to determine if we need to act further. Add a POST operation check as well, as
I'm confident we'll need that at some point.

I'm actually leaning towards doing everything in the x- header and leaving the
query string alone, because then we get the magical bucket- or object- op
processing from rgw and don't have to figure it out for ourselves.