Working ObjectStatus command.
Needed some work disabling permission checks in rgw_process_authenticated(),
and needed to crib some code from other objects to actually query the data.

Still, it returns a 200 with some XML containing a few items about the object
should it exist, and 404 if it does not exist. This is pretty much the MVP.