Initial EAK parameters support.
Set up, unconditionally for now, the EAK parameters during authentication by
mimicking relevant parts of RGWHandler_Lib::init_from_header().

In brief, for EAK we need to pull some configuration out of the request sooner
than normal so we can send more request metadata to the Authenticator.
Usually, we don't need to send EAK data, so normally we shouldn't even need to
perform this step.

Once we've agreed on the exact formatting for the EAK credential, we'll make
this setup conditional on receiving the specially-formatted credential.

Note that, as the source comment says, we're relying on some canonicalisation
that takes place in RGWREST::preprocess(), such that the bucket name is
prepended to the list of URL parameters even if it's originally used as a
domain name prefix. This saves a lot of tedious work.

Also note that it appears to me that init_from_header() redoes some work
unnecessarily - the arguments are already parsed by the time the authenticator
sees the request. I've not changed anything outside our own code.

Unit tests are due. This should not be especially hard to test.