Unbreak handoff unit test in v18.
We can't instantiate the req_state in the same way in v18. I stole the
instantiation macro DEFINE_REQ_STATE from src/test/rgw/test_rgw_lua.cc,
and then included a couple of otherwise unnecessary files in order to
not have static asserts in RGWProcessEnv when a couple of structures are
of zero length. (See comments in the commit.)

Everywhere we set up a dummy req_state, call DEFINE_REQ_STATE.