Simple unit test for GetSigningKey client side.
As with the other gRPC tests, implement a simplified server in the test in
order to test the client side that is baked into rgw.

We're only testing the client that fetches the signing key. It's hard to test
the actual chunked upload part in a unit test because it requires the whole
server. Luckily, we didn't write the part that implements chunked upload, only
the part that passes a new parameter into the AWSv4ComplMulti instance.