Unbreak WITH_JAEGER=ON compilation.
This is a bit involved, because we have out-of-tree Abseil and gRPC, and
a CMake external project (opentelemetry-cpp) which is quite hard to
configure at a distance.

- OpenTelemetry really wants to use its in-tree Abseil, and we have to turn that off.
- Add the abseil include directories before most other stuff, just to
make sure we get the right version.
- Point the otel build at our Abseil in /usr/local/abseil-cpp.
- Tell the build to run in C++17 mode or it won't work. (Abseil tries to
add newer C++ features to older compilers, so if we don't tell it we're
a newer C++ version it will compile custom versions of e.g.
std::optional and std::variant, which will break things.)
- Define HAVE_ABSEIL *everywhere*, because whenever tracer.c is included
(i.e. everywhere tracing is supported) we use HAVE_ABSEIL in a
subordinate include file to discover whether or not we need to use the
in-tree Abseil. We never want that; always define HAVE_ABSEIL.