WIP: Wire in the S3 Anonymous engine.
Duplicate rgw::auth::AnonymousEngine's authenticate() method in
rgw::auth::s3::S3AnonymousEngine. Make a few private vars of
AnonymousEngine protected to support that, and un-final
AnonymousEngine's authenticate() method.

Now, an anonymous S3 request is handled by
rgw::auth::s3::S3AnonymousEngine, and that engine has access to the
HandoffHelper shared_ptr available in the req_state.

Next step is to call out to the Authenticator if configured to do so.