WIP working anonymous authz.
By just sending empty strings to the AuthenticateRESTRequest and
teaching the Python auth prototype to understand that as 'anonymous',
we're able to authorize via the remote without effort.

However: The bucket has to be set to public to allow this! E.g. using
  aws s3api put-bucket-acl --bucket testnv --acl public-read

Without this, the internal authz checks prevent bucket access regardless
of handoff authz allowing it.